requirements.txt
chinese_calendar==1.7.2
chinesecalendar==1.7.2
Django==3.2
djangorestframework==3.13.1
pandas==1.3.5
